# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /v1/devices/{id}:
    get:
      tags:
      - device-controller
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DeviceResponseDto'
    put:
      tags:
      - device-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequestDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DeviceResponseDto'
    delete:
      tags:
      - device-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
    patch:
      tags:
      - device-controller
      operationId: updatePartially
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartialDeviceDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DeviceResponseDto'
  /v1/devices:
    get:
      tags:
      - device-controller
      operationId: getDevices
      parameters:
      - name: brand
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceResponseDto'
    post:
      tags:
      - device-controller
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequestDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DeviceResponseDto'
components:
  schemas:
    DeviceRequestDto:
      required:
      - brand
      - name
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
    DeviceResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        brand:
          type: string
        createdAt:
          type: string
          format: date-time
    PartialDeviceDto:
      required:
      - brand
      type: object
      properties:
        brand:
          type: string